+++ template-parts/content/content-search.php
@@ -12,7 +12,7 @@
 	<header class="entry-header">
 		<?php the_title( sprintf( '<h2 class="entry-title"><a href="%s" rel="bookmark">', esc_url( get_permalink() ) ), '</a></h2>' ); ?>
 
-		<?php if ( 'post' === get_post_type() || 'ndt4_news' === get_post_type() ) : ?>
+		<?php if ( 'post' === get_post_type() ) : ?>
 			<div class="meta entry-meta">
 				<?php
 				ndt4_posted_on();